Move RHF to its own directory.
