+++ gqcp/include/QCMethod/HF/RHF/CMakeLists.txt
@@ -1,0 +1,14 @@
+target_sources(gqcp
+    PRIVATE
+        DiagonalRHFFockMatrixObjective.hpp
+        RHF.hpp
+        RHFDensityMatrixCalculation.hpp
+        RHFDensityMatrixDamper.hpp
+        RHFElectronicEnergyCalculation.hpp
+        RHFErrorCalculation.hpp
+        RHFFockMatrixCalculation.hpp
+        RHFFockMatrixDiagonalization.hpp
+        RHFFockMatrixDIIS.hpp
+        RHFSCFEnvironment.hpp
+        RHFSCFSolver.hpp
+)
