+++ gqcp/include/QCMethod/HF/RHF/RHFDensityMatrixDamper.hpp
@@ -1,4 +1,4 @@
-// This file is part of GQCG-GQCP.
+// / RHF / RHF.hpps part of GQCG-GQCP.
 //
 // Copyright (C) 2017-2020  the GQCG developers
 //
@@ -20,8 +20,8 @@
 
 #include "Mathematical/Algorithm/Step.hpp"
 #include "Mathematical/Optimization/Accelerator/ConstantDamper.hpp"
-#include "QCMethod/HF/RHF.hpp"
-#include "QCMethod/HF/RHFSCFEnvironment.hpp"
+#include "QCMethod/HF/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF/RHFSCFEnvironment.hpp"
 
 
 namespace GQCP {
