+++ gqcp/include/QCMethod/HF/RHF/RHFElectronicEnergyCalculation.hpp
@@ -1,4 +1,4 @@
-// This file is part of GQCG-GQCP.
+// / RHF / RHF.hpps part of GQCG-GQCP.
 //
 // Copyright (C) 2017-2020  the GQCG developers
 //
@@ -19,21 +19,20 @@
 
 
 #include "Mathematical/Algorithm/Step.hpp"
-#include "QCMethod/HF/RHFSCFEnvironment.hpp"
-
-#include <Eigen/Dense>
+#include "QCMethod/HF/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF/RHFSCFEnvironment.hpp"
 
 
 namespace GQCP {
 
 
 /**
- *  An iteration step that solves the generalized eigenvalue problem for the current scalar/AO basis Fock matrix for the coefficient matrix.
+ *  An iteration step that calculates the current electronic RHF energy.
  * 
  *  @tparam _Scalar              the scalar type used to represent the expansion coefficient/elements of the transformation matrix
  */
 template <typename _Scalar>
-class RHFFockMatrixDiagonalization:
+class RHFElectronicEnergyCalculation:
     public Step<RHFSCFEnvironment<_Scalar>> {
 
 public:
@@ -47,21 +46,18 @@ public:
      */
 
     /**
-     *  Solve the generalized eigenvalue problem for the most recent scalar/AO Fock matrix. Add the associated coefficient matrix and orbital energies to the environment.
+     *  Calculate the current electronic RHF energy and place it in the environment
      * 
      *  @param environment              the environment that acts as a sort of calculation space
      */
     void execute(Environment& environment) override {
 
-        const auto& F = environment.fock_matrices.back();  // the most recent scalar/AO basis Fock matrix
+        const auto& D = environment.density_matrices.back();                             // the most recent density matrix
+        const ScalarSQOneElectronOperator<Scalar> F {environment.fock_matrices.back()};  // the most recent Fock matrix
+        const auto& H_core = environment.sq_hamiltonian.core();                          // the core Hamiltonian matrix
 
-        using MatrixType = Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic>;
-        Eigen::GeneralizedSelfAdjointEigenSolver<MatrixType> generalized_eigensolver(F, environment.S);
-        const TransformationMatrix<Scalar>& C = generalized_eigensolver.eigenvectors();
-        const auto& orbital_energies = generalized_eigensolver.eigenvalues();
-
-        environment.coefficient_matrices.push_back(C);
-        environment.orbital_energies.push_back(orbital_energies);
+        const auto E_electronic = QCModel::RHF<double>::calculateElectronicEnergy(D, H_core, F);
+        environment.electronic_energies.push_back(E_electronic);
     }
 };
