+++ gqcp/include/QCMethod/HF/RHF/RHFErrorCalculation.hpp
@@ -1,4 +1,4 @@
-// This file is part of GQCG-GQCP.
+// / RHF / RHF.hpps part of GQCG-GQCP.
 //
 // Copyright (C) 2017-2020  the GQCG developers
 //
@@ -19,20 +19,20 @@
 
 
 #include "Mathematical/Algorithm/Step.hpp"
-#include "QCMethod/HF/RHF.hpp"
-#include "QCMethod/HF/RHFSCFEnvironment.hpp"
+#include "QCMethod/HF/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF/RHFSCFEnvironment.hpp"
 
 
 namespace GQCP {
 
 
 /**
- *  An iteration step that calculates the current Fock matrix (expressed in the scalar/AO basis) from the current density matrix.
+ *  An iteration step that calculates the error matrix from the Fock and density matrices (expressed in the scalar/AO basis).
  * 
  *  @tparam _Scalar              the scalar type used to represent the expansion coefficient/elements of the transformation matrix
  */
 template <typename _Scalar>
-class RHFFockMatrixCalculation:
+class RHFErrorCalculation:
     public Step<RHFSCFEnvironment<_Scalar>> {
 
 public:
@@ -46,14 +46,20 @@ public:
      */
 
     /**
-     *  Calculate the current RHF density matrix (expressed in the scalar/AO basis) and place it in the environment
+     *  Calculate the current error vector and add it to the environment.
      * 
      *  @param environment              the environment that acts as a sort of calculation space
      */
     void execute(Environment& environment) override {
-        const auto& D = environment.density_matrices.back();  // the most recent density matrix
-        const auto F = QCModel::RHF<double>::calculateScalarBasisFockMatrix(D, environment.sq_hamiltonian);
-        environment.fock_matrices.push_back(F.parameters());
+
+        // Read F, D and S from the environment
+        const auto& D = environment.density_matrices.back();
+        const auto& S = environment.S;
+        const auto& F = environment.fock_matrices.back();
+
+        // Calculate the error and write it to the environment (as a vector)
+        const auto error_matrix = QCModel::RHF<Scalar>::calculateError(F, D, S);
+        environment.error_vectors.push_back(error_matrix.pairWiseReduce());
     }
 };
