+++ gqcp/include/QCMethod/HF/UHF/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(gqcp
+    PRIVATE
+        UHFSCFEnvironment.hpp
+)
