+++ gqcp/include/gqcp.hpp
@@ -159,17 +159,17 @@
 // #include "QCMethod/CI/DOCINewtonOrbitalOptimizer.hpp"
 #include "QCMethod/Geminals/AP1roGJacobiOrbitalOptimizer.hpp"
 #include "QCMethod/Geminals/AP1roGLagrangianNewtonOrbitalOptimizer.hpp"
-#include "QCMethod/HF/DiagonalRHFFockMatrixObjective.hpp"
-#include "QCMethod/HF/RHF.hpp"
-#include "QCMethod/HF/RHFDensityMatrixCalculation.hpp"
-#include "QCMethod/HF/RHFDensityMatrixDamper.hpp"
-#include "QCMethod/HF/RHFElectronicEnergyCalculation.hpp"
-#include "QCMethod/HF/RHFErrorCalculation.hpp"
-#include "QCMethod/HF/RHFFockMatrixCalculation.hpp"
-#include "QCMethod/HF/RHFFockMatrixDIIS.hpp"
-#include "QCMethod/HF/RHFFockMatrixDiagonalization.hpp"
-#include "QCMethod/HF/RHFSCFEnvironment.hpp"
-#include "QCMethod/HF/RHFSCFSolver.hpp"
+#include "QCMethod/HF/RHF/DiagonalRHFFockMatrixObjective.hpp"
+#include "QCMethod/HF/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF/RHFDensityMatrixCalculation.hpp"
+#include "QCMethod/HF/RHF/RHFDensityMatrixDamper.hpp"
+#include "QCMethod/HF/RHF/RHFElectronicEnergyCalculation.hpp"
+#include "QCMethod/HF/RHF/RHFErrorCalculation.hpp"
+#include "QCMethod/HF/RHF/RHFFockMatrixCalculation.hpp"
+#include "QCMethod/HF/RHF/RHFFockMatrixDIIS.hpp"
+#include "QCMethod/HF/RHF/RHFFockMatrixDiagonalization.hpp"
+#include "QCMethod/HF/RHF/RHFSCFEnvironment.hpp"
+#include "QCMethod/HF/RHF/RHFSCFSolver.hpp"
 #include "QCMethod/OrbitalOptimization/BaseOrbitalOptimizer.hpp"
 #include "QCMethod/OrbitalOptimization/JacobiOrbitalOptimizer.hpp"
 #include "QCMethod/OrbitalOptimization/Localization/ERJacobiLocalizer.hpp"
