+++ gqcp/tests/Processing/Properties/properties_test.cpp
@@ -1,4 +1,4 @@
-// This file is part of GQCG-GQCP.
+// / RHF / DiagonalRHFFockMatrixObjective
 //
 // Copyright (C) 2017-2020  the GQCG developers
 //
@@ -26,9 +26,9 @@
 #include "Processing/Properties/properties.hpp"
 #include "Processing/RDM/RDMCalculator.hpp"
 #include "QCMethod/CI/HamiltonianBuilder/FCI.hpp"
-#include "QCMethod/HF/DiagonalRHFFockMatrixObjective.hpp"
-#include "QCMethod/HF/RHF.hpp"
-#include "QCMethod/HF/RHFSCFSolver.hpp"
+#include "QCMethod/HF/RHF/DiagonalRHFFockMatrixObjective.hpp"
+#include "QCMethod/HF/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF/RHFSCFSolver.hpp"
 #include "Utilities/units.hpp"
