+++ gqcp/tests/QCMethod/Geminals/QCMethod_vAP1roG_test.cpp
@@ -1,4 +1,4 @@
-// This file is part of GQCG-GQCP.
+// / RHF / DiagonalRHFFockMatrixObjective
 //
 // Copyright (C) 2017-2020  the GQCG developers
 //
@@ -25,9 +25,9 @@
 #include "Processing/Properties/expectation_values.hpp"
 #include "QCMethod/Geminals/PSEnvironment.hpp"
 #include "QCMethod/Geminals/vAP1roG.hpp"
-#include "QCMethod/HF/DiagonalRHFFockMatrixObjective.hpp"
-#include "QCMethod/HF/RHF.hpp"
-#include "QCMethod/HF/RHFSCFSolver.hpp"
+#include "QCMethod/HF/RHF/DiagonalRHFFockMatrixObjective.hpp"
+#include "QCMethod/HF/RHF/RHF.hpp"
+#include "QCMethod/HF/RHF/RHFSCFSolver.hpp"
 #include "QCModel/Geminals/vAP1roG.hpp"
