+++ gqcp/tests/QCMethod/HF/CMakeLists.txt
@@ -1,6 +1,3 @@
-list(APPEND test_target_sources
-    ${CMAKE_CURRENT_SOURCE_DIR}/QCMethod_RHF_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/RHFSCFEnvironment_test.cpp
-)
+add_subdirectory(RHF)
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
