+++ gqcpy/src/QCMethod/HF/RHF/RHFSCFSolver.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "QCMethod/HF/RHFSCFSolver.hpp"
+#include "QCMethod/HF/RHF/RHFSCFSolver.hpp"
 
 #include <pybind11/pybind11.h>
